pyproject.toml
[project]
name = "social-media-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
